+++ ydb/library/workload/benchmark_base/workload.h
@@ -43,7 +43,7 @@ public:
 class TWorkloadGeneratorBase : public IWorkloadQueryGenerator {
 public:
     explicit TWorkloadGeneratorBase(const TWorkloadBaseParams& params);
-    void Init() override {};
+    void Init() override final {};
     std::string GetDDLQueries() const override final;
     TVector<std::string> GetCleanPaths() const override final;