+++ ydb/apps/ydb/CHANGELOG.md
@@ -1,4 +1,4 @@
-* Added `auto` as a supported value for the `--float-mode` option in the `ydb workload tpch run` and `ydb workload tpcds run` commands.
+* Removed `--float-mode` option from the `ydb workload tpch run` and `ydb workload tpcds run` commands. Now float mode is detected automatically.
 * Added final execute statistics to `ydb workload * run` commands.
 * Fixed a bug where the `ydb import file csv command` with the `--newline-delimited` option could get stuck if the input had incorrect data.
 * Fixed a bug with the progress bar display in the `ydb workload clickbench import files` command — incorrect percentage value and excessive line breaks causing duplicated progress lines.